NONE: Improve naming for use cases
